// infrastructure/client-library/src/types.ts
export enum TransactionType {
  Send = "send",
  Receive = "receive",
  Trade = "trade",
  List = "list",
  Claim = "claim",
}

export enum AssetType {
  BTC = "btc",
  BRC20 = "brc-20",
  RUNE = "rune",
  ALKANE = "alkane",
  COLLECTIBLE = "collectible",
}

export enum BRC20Operation {
  Deploy = "deploy",
  Mint = "mint",
  Transfer = "transfer",
}

export enum RuneOperation {
  Etching = "etching",
  Mint = "mint",
  Transfer = "transfer",
}

export interface TransactionIntent extends BaseIntent {
  type: IntentType.Transaction;
  transactionType: TransactionType;
  assetType: AssetType;
  transactionIds: string[];
  btcAmount: number;
}

export interface BTCTransactionIntent extends TransactionIntent {
  assetType: AssetType.BTC;
  amount: number;
}

export interface BRC20TransactionIntent extends TransactionIntent {
  assetType: AssetType.BRC20;
  ticker: string;
  tickerAmount?: number;
  operation: BRC20Operation;
  max?: number;
  limit?: number;
}

export interface RuneEtchingTransactionIntent extends TransactionIntent {
  assetType: AssetType.RUNE;
  operation: RuneOperation.Etching;
  runeName: string;
  inscription?: CategorizedInscription;
}

export interface RuneMintTransactionIntent extends TransactionIntent {
  assetType: AssetType.RUNE;
  operation: RuneOperation.Mint;
  runeId: string;
  runeName: string;
  runeAmount: bigint;
  runeDivisibility: number;
  inscription?: CategorizedInscription;
}

export interface RuneTransferTransactionIntent extends TransactionIntent {
  assetType: AssetType.RUNE;
  operation: RuneOperation.Transfer;
  runeId: string;
  runeName: string;
  runeAmount: bigint;
  runeDivisibility: number;
  inscription?: CategorizedInscription;
}

export type RuneTransactionIntent =
  | RuneEtchingTransactionIntent
  | RuneMintTransactionIntent
  | RuneTransferTransactionIntent;

export interface CollectibleTransactionIntent extends TransactionIntent {
  assetType: AssetType.COLLECTIBLE;
  inscriptionId: string;
  contentType: string;
  content: string;
  receiverAddress?: string;
}

export interface BRC20TradeTransactionIntent extends TransactionIntent {
  assetType: AssetType.BRC20;
  transactionType: TransactionType.Trade;
  ticker: string;
  tickerAmount: number;
  totalPrice: number;
}

export interface RuneTradeTransactionIntent extends TransactionIntent {
  assetType: AssetType.RUNE;
  transactionType: TransactionType.Trade;
  operation: RuneOperation.Transfer;
  runeName: string;
  runeAmount: bigint;
  totalPrice: number;
}

export interface CollectibleTradeTransactionIntent extends TransactionIntent {
  assetType: AssetType.COLLECTIBLE;
  transactionType: TransactionType.Trade;
  inscriptionId: string;
  contentType: string;
  content: string;
  totalPrice: number;
}

export interface CollectibleListTransactionIntent extends TransactionIntent {
  assetType: AssetType.COLLECTIBLE;
  transactionType: TransactionType.List;
  marketplace: string;
  inscriptionId: string;
  collectionName: string;
  inscriptionName: string;
  price: number;
  listingId?: string;
}

export interface CollectibleClaimTransactionIntent extends TransactionIntent {
  assetType: AssetType.COLLECTIBLE;
  transactionType: TransactionType.Claim;
  inscriptionId?: string;
  imageUrl: string;
  collectionName: string;
}

export type WalletIntent =
  | BTCTransactionIntent
  | BRC20TransactionIntent
  | RuneTransactionIntent
  | CollectibleTransactionIntent
  | BRC20TradeTransactionIntent
  | RuneTradeTransactionIntent
  | CollectibleTradeTransactionIntent
  | CollectibleListTransactionIntent
  | CollectibleClaimTransactionIntent;
// ...
export interface BackendServiceConfig {
  url: string;
  wsUrl?: string;
}